+++ src/UglyToad.PdfPig.Tests/Integration/ColorSpaceCacheTests.cs
@@ -1,20 +1,14 @@
-namespace UglyToad.PdfPig.Tests.Integration
+﻿namespace UglyToad.PdfPig.Tests.Integration
 {
     using System.Collections.Generic;
     using System.Linq;
     using PdfPig.Content;
     using PdfPig.Filters;
     using PdfPig.Graphics.Colors;
-    using PdfPig.PdfFonts;
     using PdfPig.Tokens;
 
     public class ColorSpaceCacheTests
     {
-        private sealed class NoOpFontFactory : IFontFactory
-        {
-            public IFont Get(DictionaryToken dictionary) => null!;
-        }
-
         [Fact]
         public void ShadingsSharingColorSpaceObjectShareOneInstance()
         {
